assing static videos page json

+++ src/services/videosService.js
@@ -1,15 +1,32 @@
+const base = import.meta.env.BASE_URL;
+
 const API = "http://localhost:3001/videos";
+
+const isLocal =
+  window.location.hostname === "localhost" ||
+  window.location.hostname === "127.0.0.1";
 
 
 // GET ALL
 export async function getVideos() {
-  const res = await fetch(API);
+
+  if (isLocal) {
+    const res = await fetch(API);
+    return res.json();
+  }
+
+  const res = await fetch(`${base}videos/videos.json`);
   return res.json();
 }
 
 
 // CREATE
 export async function createVideo(data) {
+
+  if (!isLocal) {
+    throw new Error("Criação disponível apenas no localhost.");
+  }
+
   const res = await fetch(API, {
     method: "POST",
     headers: {
@@ -24,6 +41,11 @@ export async function createVideo(data) {
 
 // UPDATE
 export async function updateVideo(id, data) {
+
+  if (!isLocal) {
+    throw new Error("Edição disponível apenas no localhost.");
+  }
+
   const res = await fetch(`${API}/${id}`, {
     method: "PUT",
     headers: {
@@ -38,6 +60,11 @@ export async function updateVideo(id, data) {
 
 // DELETE
 export async function deleteVideo(id) {
+
+  if (!isLocal) {
+    throw new Error("Exclusão disponível apenas no localhost.");
+  }
+
   await fetch(`${API}/${id}`, {
     method: "DELETE"
   });
